requirements.txt
charset-normalizer==2.0.12
idna==3.3
imageio==2.16.1
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.0.1
requests==2.27.1
urllib3==1.26.9
